RAC-107 refactor : 대학원생 인증 상태 enum사용 및 3단계 분할
