+++ src/main/java/com/postgraduate/domain/senior/application/dto/res/SeniorInfoResponse.java
@@ -1,5 +1,6 @@
 package com.postgraduate.domain.senior.application.dto.res;
 
+import com.postgraduate.domain.senior.domain.entity.constant.Status;
 import lombok.AllArgsConstructor;
 import lombok.Builder;
 import lombok.Getter;
@@ -10,6 +11,6 @@ import lombok.Getter;
 public class SeniorInfoResponse {
     private String nickName;
     private String profile;
-    private boolean certificationRegister;
+    private Status certificationRegister;
     private boolean profileRegister;
 }
