+++ src/main/java/com/postgraduate/domain/senior/domain/entity/Senior.java
@@ -1,6 +1,7 @@
 package com.postgraduate.domain.senior.domain.entity;
 
 import com.postgraduate.domain.senior.application.dto.req.SeniorProfileRequest;
+import com.postgraduate.domain.senior.domain.entity.constant.Status;
 import com.postgraduate.domain.user.domain.entity.User;
 import jakarta.persistence.*;
 import lombok.AllArgsConstructor;
@@ -8,6 +9,9 @@ import lombok.Builder;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
 import org.hibernate.annotations.DynamicInsert;
+
+import static com.postgraduate.domain.senior.domain.entity.constant.Status.NOT_APPROVE;
+import static com.postgraduate.domain.senior.domain.entity.constant.Status.WAITING;
 
 @Entity
 @Builder
@@ -30,8 +34,9 @@ public class Senior {
     private String rrn;
 
     @Column(nullable = false)
+    @Enumerated(EnumType.STRING)
     @Builder.Default
-    private Boolean status = false;
+    private Status status = WAITING;
 
     @Column(nullable = false)
     private int hit;
@@ -50,6 +55,6 @@ public class Senior {
 
     public void updateCertification(String certification) {
         this.certification = certification;
-        this.status = false;
+        this.status = WAITING;
     }
 }
