+++ src/main/java/com/postgraduate/domain/senior/domain/entity/constant/Status.java
@@ -1,0 +1,10 @@
+package com.postgraduate.domain.senior.domain.entity.constant;
+
+import lombok.Getter;
+import lombok.RequiredArgsConstructor;
+
+@RequiredArgsConstructor
+@Getter
+public enum Status {
+    NOT_APPROVE, WAITING, APPROVE
+}
