+++ src/main/java/com/postgraduate/domain/senior/domain/repository/SeniorRepository.java
@@ -1,4 +1,4 @@
-package com.postgraduate.domain.senior.domain.entity.repository;
+package com.postgraduate.domain.senior.domain.repository;
 
 import com.postgraduate.domain.senior.domain.entity.Senior;
 import com.postgraduate.domain.user.domain.entity.User;
