+++ src/main/java/com/postgraduate/domain/senior/domain/service/SeniorGetService.java
@@ -1,7 +1,7 @@
 package com.postgraduate.domain.senior.domain.service;
 
 import com.postgraduate.domain.senior.domain.entity.Senior;
-import com.postgraduate.domain.senior.domain.entity.repository.SeniorRepository;
+import com.postgraduate.domain.senior.domain.repository.SeniorRepository;
 import com.postgraduate.domain.user.domain.entity.User;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
